dataprep/requirements-2.txt
google-images-download==2.8.0
numpy==1.17.4
Pillow==6.2.1
selenium==3.141.0
urllib3==1.25.7
